Extract command handler out of the registerCommand callback
All the command's control flow lived in an anonymous callback passed
to registerCommand, which made it impossible to unit test the handler
itself (only the pure helpers extracted earlier could be tested).

Move the handler into a named toggleTestFile() function in its own
module, so activate() is now just wiring.
