+++ src/extension.ts
@@ -1,40 +1,8 @@
 import * as vscode from 'vscode';
-import { basename } from 'path';
-import { toggleFileName, extractPackage } from './logic';
+import { toggleTestFile } from './toggleTestFile';
 
 export function activate(context: vscode.ExtensionContext) {
-  const disposable = vscode.commands.registerCommand('vscode-quick-junit.toggleTestFile', async () => {
-    const editor = vscode.window.activeTextEditor;
-    if (!editor) {
-      vscode.window.showInformationMessage('No file is currently open.');
-      return;
-    }
-
-    const currentFileName = basename(editor.document.uri.fsPath);
-    const targetFileName = toggleFileName(currentFileName);
-
-    const files = await vscode.workspace.findFiles(`**/${targetFileName}`);
-    if (files.length === 0) {
-      return;
-    }
-
-    if (files.length === 1) {
-      const document = await vscode.workspace.openTextDocument(files[0]);
-      await vscode.window.showTextDocument(document);
-      return;
-    }
-
-    const currentPackage = extractPackage(editor.document.getText());
-    for (const file of files) {
-      const document = await vscode.workspace.openTextDocument(file);
-      const candidatePackage = extractPackage(document.getText());
-      if (currentPackage === candidatePackage) {
-        await vscode.window.showTextDocument(document);
-        return;
-      }
-    }
-  });
-
+  const disposable = vscode.commands.registerCommand('vscode-quick-junit.toggleTestFile', toggleTestFile);
   context.subscriptions.push(disposable);
 }
